Revert "various changes for supporting existing logic in custom modules.  more still to come"
This reverts commit 481c7c8edf0d8fcbc2e83fdc9d38e43753343d05.
