+++ android/runtime/common/src/java/org/appcelerator/kroll/KrollInvocation.java
@@ -1,0 +1,22 @@
+/**
+ * Appcelerator Titanium Mobile
+ * Copyright (c) 2011 by Appcelerator, Inc. All Rights Reserved.
+ * Licensed under the terms of the Apache Public License
+ * Please see the LICENSE included with this distribution for details.
+ */
+package org.appcelerator.kroll;
+
+public class KrollInvocation
+{
+	private String sourceUrl;
+
+	public KrollInvocation(String sourceUrl)
+	{
+		this.sourceUrl = sourceUrl;
+	}
+
+	public String getSourceUrl()
+	{
+		return sourceUrl;
+	}
+}
